% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cojo.R
\name{cojo}
\alias{cojo}
\title{Run GCTA-COJO.}
\usage{
cojo(
  tab,
  bed,
  method = c("select", "cond"),
  cmd = "",
  snps_cond = NULL,
  dir_run = "run_cojo",
  tool = getOption("finemapr_cojo"),
  args = ""
)
}
\description{
http://gcta.freeforums.net/thread/178/conditional-joint-analysis-using-summary
}
\details{
tab <- "~/git/hemostat/scc/results/sumstats-cojo/sumstats-1.cojo"
bed <- "~/git/hemostat/scc/data/FineMapping/forGCTA_onco_r1.bed"
}
