% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapr.R
\name{finemapr}
\alias{finemapr}
\title{Run FINEMAP.}
\usage{
finemapr(
  tab,
  ld,
  n,
  annot,
  annotations,
  prop_credible = 0.95,
  method = c("finemap", "paintor"),
  dir_run,
  tool,
  args = "",
  prior_k,
  save_ld = FALSE,
  ret = c("results", "zscore", "ld")
)
}
\description{
Run FINEMAP.
}
\examples{
\dontrun{
ex <- example_finemap()
out <- finemapr(
  list(ex$tab1, ex$tab2), 
  list(ex$ld1, ex$ld2), 
  list(ex$n1, ex$n2), 
  args = "--n-causal-max 1")
out <- finemapr(list(ex$tab1, ex$tab2), list(ex$ld1, ex$ld2),
  list(ex$n1, ex$n2), method = "paintor", args = "-enumerate 1")
}
}
