% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R, R/classMethods.R, R/finemapr.R
\name{plot_zscore.Finemapr}
\alias{plot_zscore.Finemapr}
\alias{Finemapr}
\alias{plot_ncausal}
\alias{print.Finemapr}
\alias{plot.Finemapr}
\alias{plot_snp.Finemapr}
\alias{extract_credible_set.Finemapr}
\alias{process_tab.Finemapr}
\alias{process_ld.Finemapr}
\title{S3 class Finemapr.}
\usage{
\method{plot_zscore}{Finemapr}(
  x,
  locus = 1,
  label_size = getOption("finemapr_label_size"),
  selected,
  force = 20,
  color_main = "grey75",
  color_selected = "aquamarine4",
  color_credible = "dodgerblue4",
  color_missing = "brown",
  plot_missing = FALSE,
  ...
)

plot_ncausal(x, ...)

\method{print}{Finemapr}(x, ...)

\method{print}{Finemapr}(x, ...)

\method{plot}{Finemapr}(x, ...)

\method{plot_snp}{Finemapr}(
  x,
  lim_prob = c(0, 1.5),
  label_size = getOption("finemapr_label_size"),
  top_rank = getOption("top_rank"),
  ...
)

\method{extract_credible_set}{Finemapr}(x, ...)

\method{process_tab}{Finemapr}(x, tabs, ...)

\method{process_ld}{Finemapr}(x, lds, ...)
}
\description{
S3 class Finemapr.
}
\note{
Colors: http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf
}
