% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R, R/classFinemaprFinemap.R
\name{filename_zscore.FinemaprFinemap}
\alias{filename_zscore.FinemaprFinemap}
\alias{filename_ld.FinemaprFinemap}
\alias{filename_snp.FinemaprFinemap}
\alias{filename_config.FinemaprFinemap}
\alias{filename_k.FinemaprFinemap}
\alias{filename_log.FinemaprFinemap}
\alias{filename_master.FinemaprFinemap}
\alias{FinemaprFinemap}
\alias{write_files.FinemaprFinemap}
\alias{run_tool.FinemaprFinemap}
\alias{collect_results.FinemaprFinemap}
\alias{print.FinemaprFinemap}
\alias{plot.FinemaprFinemap}
\alias{plot_ncausal.FinemaprFinemap}
\alias{plot_config.FinemaprFinemap}
\alias{plot_snp.FinemaprFinemap}
\title{S3 class FinemaprFinemap.}
\usage{
\method{filename_zscore}{FinemaprFinemap}(x, locus = 1)

\method{filename_ld}{FinemaprFinemap}(x, locus = 1)

\method{filename_snp}{FinemaprFinemap}(x, locus)

\method{filename_config}{FinemaprFinemap}(x, locus)

\method{filename_k}{FinemaprFinemap}(x, locus)

\method{filename_log}{FinemaprFinemap}(x, locus)

\method{filename_master}{FinemaprFinemap}(x)

\method{write_files}{FinemaprFinemap}(x, ...)

\method{run_tool}{FinemaprFinemap}(x, ...)

\method{collect_results}{FinemaprFinemap}(x, ...)

\method{print}{FinemaprFinemap}(x, ...)

\method{plot}{FinemaprFinemap}(
  x,
  grid_nrow = NULL,
  grid_ncol = NULL,
  label_size = getOption("finemapr_label_size"),
  label_size_config = label_size,
  label_size_snp = label_size,
  top_rank = getOption("top_rank"),
  top_rank_config = top_rank,
  top_rank_snp = top_rank,
  lim_prob = getOption("lim_prob"),
  lim_prob_config = lim_prob,
  lim_prob_snp = lim_prob,
  lim_prob_ncausal = lim_prob,
  ...
)

\method{plot_ncausal}{FinemaprFinemap}(x, locus = 1, lim_prob = c(0, 1), ...)

\method{plot_config}{FinemaprFinemap}(
  x,
  locus = 1,
  lim_prob = c(0, 1.5),
  label_size = getOption("finemapr_label_size"),
  top_rank = getOption("top_rank"),
  ...
)

\method{plot_snp}{FinemaprFinemap}(
  x,
  locus = 1,
  lim_prob = c(0, 1.5),
  label_size = getOption("finemapr_label_size"),
  top_rank = getOption("top_rank"),
  ...
)
}
\description{
S3 class FinemaprFinemap.
}
