% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{proxysnps_ref_ld}
\alias{proxysnps_ref_ld}
\title{https://github.com/slowkow/proxysnps/blob/master/R/get_vcf.R}
\usage{
proxysnps_ref_ld(
  vcf,
  tol = 1e-10,
  measure = c("R.squared", "D.prime"),
  verbose = 0
)
}
\description{
https://github.com/slowkow/proxysnps/blob/master/R/get_vcf.R
}
