% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{proxysnps_ref_vcf}
\alias{proxysnps_ref_vcf}
\title{https://github.com/slowkow/proxysnps/blob/master/R/get_vcf.R}
\usage{
proxysnps_ref_vcf(
  chr = 16,
  start = 53767000,
  end = 53768000,
  pop = "EUR",
  snps = NULL,
  verbose = 0
)
}
\description{
https://github.com/slowkow/proxysnps/blob/master/R/get_vcf.R
}
