% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ld}
\alias{read_ld}
\title{Read LD matrix from a file.}
\usage{
read_ld(file_ld, snps)
}
\description{
Read LD matrix from a file.
}
\examples{
f1 <- system.file("extdata/region1.ld", package = "finemapr")
ld <- read_ld(f1)

}
