% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_zscore}
\alias{read_zscore}
\title{Read z-scores from a file.}
\usage{
read_zscore(file_zscore)
}
\description{
Read z-scores from a file.
}
\examples{
f1 <- system.file("extdata/region1.z", package = "finemapr")
ztab <- read_zscore(f1)

}
