% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caviar.R
\name{run_caviar}
\alias{run_caviar}
\title{Run CAVIAR.}
\usage{
run_caviar(
  tab,
  ld,
  dir_run = "run_caviar",
  tool = getOption("finemapr_caviar"),
  args = ""
)
}
\description{
Output file (`*_post`): 
 column #1 is the variant name;
 column #2 is the posterior prob. that the variant is causal
   (https://github.com/fhormoz/caviar/issues/1#issuecomment-286521771);
 column #3 is the amount that this variant contributes to 
   95%-causal credible set.
}
\examples{
\dontrun{
ex <- example_finemap()
out <- run_caviar(ex$tab1, ex$ld1, args = "-c 2")
out <- run_caviar(ex$tab1, ex$ld1, args = "-c 0")
}
}
