% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{sumstats_bmi}
\alias{sumstats_bmi}
\title{https://data.broadinstitute.org/alkesgroup/sumstats_formatted/readme.txt
https://github.com/bulik/ldsc/wiki/Partitioned-Heritability}
\usage{
sumstats_bmi()
}
\description{
https://data.broadinstitute.org/alkesgroup/sumstats_formatted/readme.txt
https://github.com/bulik/ldsc/wiki/Partitioned-Heritability
}
