% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf.R
\name{abf}
\alias{abf}
\title{abf, needs title}
\usage{
abf(effect, se, snp = paste0("snp", seq(1, nrow(effect))))
}
\description{
abf, needs title
}
\examples{
if (interactive()) {
 ex <- example_finemap()
 se <- rep(1/sqrt(ex$n1), nrow(ex$tab1))
 effect <- ex$tab1$zscore * se
 out <- abf(effect, se)
}
}
