% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapr.R
\name{finemapr}
\alias{finemapr}
\title{Run FINEMAP.}
\usage{
finemapr(tab, ld, n, method = c("finemap"), dir_run,
  tool = getOption("finemapr_finemap"), args = "")
}
\description{
Run FINEMAP.
}
\examples{
ex <- example_finemap()
out <- finemapr(ex$tab1, ex$ld1, ex$n1, args = "--n-causal-max 1")

}
