% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{Finemapr}
\alias{Finemapr}
\alias{plot_ncausal}
\alias{print.Finemapr}
\alias{plot.Finemapr}
\alias{plot_snp.Finemapr}
\title{S3 class Finemapr.}
\usage{
plot_ncausal(x, ...)

\method{print}{Finemapr}(x, ...)

\method{plot}{Finemapr}(x, ...)

\method{plot_snp}{Finemapr}(x, lim_prob = c(0, 1.5),
  label_size = getOption("finemapr_label_size"),
  top_rank = getOption("top_rank"), ...)
}
\description{
S3 class Finemapr.
}
