% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R, R/classMethods.R, R/finemapr.R
\name{plot_zscore.Finemapr}
\alias{plot_zscore.Finemapr}
\alias{Finemapr}
\alias{plot_ncausal}
\alias{print.Finemapr}
\alias{plot.Finemapr}
\alias{plot_snp.Finemapr}
\alias{process_tab.Finemapr}
\alias{process_ld.Finemapr}
\title{S3 class Finemapr.}
\usage{
\method{plot_zscore}{Finemapr}(x, locus = 1,
  label_size = getOption("finemapr_label_size"), selected, force = 20,
  color_selected = "green", color_credible = "blue", ...)

plot_ncausal(x, ...)

print.Finemapr(x, ...)

\method{plot}{Finemapr}(x, ...)

\method{plot_snp}{Finemapr}(x, lim_prob = c(0, 1.5),
  label_size = getOption("finemapr_label_size"),
  top_rank = getOption("top_rank"), ...)

\method{process_tab}{Finemapr}(x, tabs, ...)

\method{process_ld}{Finemapr}(x, lds, ...)
}
\description{
S3 class Finemapr.
}
