% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFinemaprCaviar.R
\name{FinemaprCaviar}
\alias{FinemaprCaviar}
\alias{print.FinemaprCaviar}
\alias{plot.FinemaprCaviar}
\alias{plot_snp.FinemaprCaviar}
\title{S3 class FinemaprCaviar.}
\usage{
print.FinemaprCaviar(x, ...)

\method{plot}{FinemaprCaviar}(x, ...)

\method{plot_snp}{FinemaprCaviar}(x, lim_prob = c(0, 1.5),
  label_size = getOption("finemapr_label_size"),
  top_rank = getOption("top_rank"), ...)
}
\description{
S3 class FinemaprCaviar.
}
