% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFinemaprFinemap.R
\name{FinemaprFinemap}
\alias{FinemaprFinemap}
\alias{print.FinemaprFinemap}
\alias{plot.FinemaprFinemap}
\title{S3 class FinemaprFinemap.}
\usage{
\method{print}{FinemaprFinemap}(x, ...)

\method{plot}{FinemaprFinemap}(x, ...)
}
\description{
S3 class FinemaprFinemap.
}
