% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFinemaprPaintor.R
\name{FinemaprPaintor}
\alias{FinemaprPaintor}
\alias{print.FinemaprPaintor}
\title{S3 class FinemaprPaintor.}
\usage{
\method{print}{FinemaprPaintor}(x, ...)
}
\description{
S3 class FinemaprPaintor.
}
