% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caviar.R
\name{run_caviar}
\alias{run_caviar}
\title{Run CAVIAR.}
\usage{
run_caviar(tab, ld, dir_run = "run_caviar",
  tool = getOption("finemapr_caviar"), args = "")
}
\description{
Run CAVIAR.
}
\examples{
ex <- example_finemap()
out <- run_caviar(ex$tab1, ex$ld1, args = "-c 2")
out <- run_caviar(ex$tab1, ex$ld1, args = "-c 0")

}
