% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap.R
\name{run_finemap}
\alias{run_finemap}
\title{Run FINEMAP.}
\usage{
run_finemap(
  tab,
  ld,
  n,
  dir_run = "run_finemap",
  tool = getOption("finemapr_finemap"),
  args = ""
)
}
\description{
Run FINEMAP.
}
\examples{
ex <- example_finemap()
out <- run_finemap(ex$tab1, ex$ld1, ex$n1, args = "--n-causal-max 3")
out <- run_finemap(ex$tab1, ex$ld1, ex$n1, args = "--n-causal-max 1")

}
