% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintor.R
\name{run_paintor}
\alias{run_paintor}
\title{Run Paintor.}
\usage{
run_paintor(
  tab,
  ld,
  n,
  annot,
  annotations,
  dir_run = "run_paintor",
  tool = getOption("finemapr_paintor"),
  args = ""
)
}
\description{
Run Paintor.
}
\examples{
ex <- example_finemap()
out <- run_paintor(ex$tab1, ex$ld1)

}
