% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintor.R
\name{run_paintor}
\alias{run_paintor}
\title{Run Paintor.}
\usage{
run_paintor(
  tab,
  ld,
  n,
  annot,
  annotations,
  dir_run = "run_paintor",
  tool = getOption("finemapr_paintor"),
  args = ""
)
}
\description{
Run Paintor.
}
\examples{
if (interactive()) { # doesn't seem to work with PAINTOR 3.0
 ex <- example_finemap()   # call needs additional parameters like -mcmc ... and it writes in your current folder!
 if (file.exists("paintor")) {
 out <- run_paintor(ex$tab1, ex$ld1)
 }
}
}
